chore: removing redundant classes, methods, and files; moving files around to more suitable directories; updating all references
