+++ src/App/ChaosRecipeEnhancer.UI/ChaosRecipeEnhancer.UI.csproj
@@ -61,7 +61,6 @@
 			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
 		</None>
 		<!-- Documentation -->
-		<Content Include="Extensions\README.md" />
 	</ItemGroup>
 	<ItemGroup>
 		<Compile Update="Properties\Resources.Designer.cs">
