--- src/App/ChaosRecipeEnhancer.UI/Extensions/ObjectExtensions.cs
@@ -1,9 +1,0 @@
-﻿namespace ChaosRecipeEnhancer.UI.Extensions;
-
-public static class ObjectExtensions
-{
-	public static object GetPropertyValue(object src, string propertyName)
-	{
-		return src.GetType().GetProperty(propertyName)?.GetValue(src);
-	}
-}
