--- src/App/ChaosRecipeEnhancer.UI/Extensions/README.md
@@ -1,7 +1,0 @@
-﻿# `EnhancePoEApp/ChaosRecipeEnhancer.UI/Extensions`
-
-Any classes and methods not belonging to a particular piece of business logic of app UI belong here.
-
-The idea is to split the different extensions by domain (i.e. classes and methods that make it easier to interact with native Windows functionality).
-
-If a piece of 'logical' code coincides with a specific business process, consider moving it to `BusinessLogic`.
