+++ src/App/ChaosRecipeEnhancer.UI/Services/FilterManipulation/ReloadFilterService.cs
@@ -2,8 +2,8 @@
 using System.IO;
 using System.Linq;
 using System.Windows.Forms;
-using ChaosRecipeEnhancer.UI.Extensions.Native;
 using ChaosRecipeEnhancer.UI.Properties;
+using ChaosRecipeEnhancer.UI.Utilities.Native;
 using ChaosRecipeEnhancer.UI.Windows;
 
 namespace ChaosRecipeEnhancer.UI.Services.FilterManipulation;
@@ -21,7 +21,7 @@ public class ReloadFilterService : IReloadFilterService
         if (chatCommand is null) return;
 
         // Map all current window names to their associated "handle to a window" pointers (HWND)
-        var openWindows = NativeWindowExtensions.GetOpenWindows();
+        var openWindows = WindowsUtilitiesForOverlays.GetOpenWindows();
 
         foreach (var window in openWindows)
         {
@@ -34,7 +34,7 @@ public class ReloadFilterService : IReloadFilterService
         // Find the Process ID associated with the 'Path of Exile' game window
         var poeWindow = openWindows.FirstOrDefault(x => x.Value == "Path of Exile").Key;
 
-        if (NativeWindowExtensions.CheckIfWindowExists(poeWindow))
+        if (WindowsUtilitiesForOverlays.CheckIfWindowExists(poeWindow))
         {
             ErrorWindow.Spawn(
                 "Could not find PoE window! Please make sure PoE is running.\n\n" +
@@ -46,7 +46,7 @@ public class ReloadFilterService : IReloadFilterService
         }
 
         // Get 'Path of Exile' window in the foreground to actually send input to said window
-        NativeWindowExtensions.SetForegroundWindow(poeWindow);
+        WindowsUtilitiesForOverlays.SetForegroundWindow(poeWindow);
 
         // Always clear the clipboard first
         Clipboard.Clear();
