+++ src/App/ChaosRecipeEnhancer.UI/Utilities/Native/WindowsUtilitiesForOverlays.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Runtime.InteropServices;
+using System.Text;
 using System.Windows;
 using System.Windows.Interop;
 
@@ -11,11 +13,29 @@ internal static class WindowsUtilitiesForOverlays
     public const int WS_EX_TOOLWINDOW = 0x00000080;
     public const int GWL_EXSTYLE = -20;
 
-    [DllImport("user32.dll")]
-    public static extern int GetWindowLong(IntPtr hwnd, int index);
+    [DllImport("USER32.dll")]
+    private static extern int GetWindowLong(IntPtr hWnd, int index);
 
-    [DllImport("user32.dll")]
-    public static extern int SetWindowLong(IntPtr hwnd, int index, int newStyle);
+    [DllImport("USER32.dll")]
+    private static extern int SetWindowLong(IntPtr hWnd, int index, int newStyle);
+
+    [DllImport("USER32.DLL")]
+    private static extern bool EnumWindows(EnumWindowsProc enumFunc, int lParam);
+
+    [DllImport("USER32.DLL")]
+    private static extern int GetWindowText(IntPtr hWnd, StringBuilder lpString, int nMaxCount);
+
+    [DllImport("USER32.DLL")]
+    private static extern int GetWindowTextLength(IntPtr hWnd);
+
+    [DllImport("USER32.DLL")]
+    private static extern bool IsWindowVisible(IntPtr hWnd);
+
+    [DllImport("USER32.DLL")]
+    private static extern IntPtr GetShellWindow();
+
+    [DllImport("USER32.DLL", CharSet = CharSet.Auto, SetLastError = true)]
+    public static extern bool SetForegroundWindow(IntPtr hWnd);
 
     public static void MakeTransparent(Window window)
     {
@@ -38,4 +58,37 @@ internal static class WindowsUtilitiesForOverlays
         int extendedStyle = GetWindowLong(hwnd, GWL_EXSTYLE);
         _ = SetWindowLong(hwnd, GWL_EXSTYLE, extendedStyle | WS_EX_TOOLWINDOW);
     }
+
+    /// <summary>Returns a dictionary that contains the handle and title of all the open windows.</summary>
+    /// <returns>A dictionary that contains the handle and title of all the open windows.</returns>
+    /// REF: https://stackoverflow.com/a/43640787
+    public static IDictionary<IntPtr, string> GetOpenWindows()
+    {
+        var shellWindow = GetShellWindow();
+        var windows = new Dictionary<IntPtr, string>();
+
+        EnumWindows(delegate (IntPtr hWnd, int lParam)
+        {
+            if (hWnd == shellWindow) return true;
+            if (!IsWindowVisible(hWnd)) return true;
+
+            var length = GetWindowTextLength(hWnd);
+            if (length == 0) return true;
+
+            var builder = new StringBuilder(length);
+            GetWindowText(hWnd, builder, length + 1);
+
+            windows[hWnd] = builder.ToString();
+            return true;
+        }, 0);
+
+        return windows;
+    }
+
+    public static bool CheckIfWindowExists(IntPtr window)
+    {
+        return window == IntPtr.Zero;
+    }
+
+    private delegate bool EnumWindowsProc(IntPtr hWnd, int lParam);
 }
