+++ src/App/ChaosRecipeEnhancer.UI/Utilities/ZemotoCommon/ViewModelBase.cs
@@ -1,5 +1,4 @@
 ﻿using ChaosRecipeEnhancer.UI.Properties;
-using CommunityToolkit.Mvvm.ComponentModel;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
